mypy: fix no-any-return/arg-type/var-annotated errors in core/metrics/models/preprocessing/reporting/utils (recovered)
Wrapped bare numpy-arithmetic returns (division, indexing) that mypy
stubs type as Any in np.asarray(...)/bool(...)/float(...) across
core/, metrics/, models/ensembling, preprocessing/, reporting/charts,
reporting/renderers, signal/, utils/. Param defaults of `X: T = None`
retyped to `Optional[T] = None` where flagged.

+++ src/mlframe/core/set_similarity.py
@@ -86,7 +86,7 @@ def kulczynski(a, b) -> float:
         return 1.0
     if na == 0.0 or nb == 0.0:
         return 0.0
-    return 0.5 * inter * (1.0 / na + 1.0 / nb)
+    return float(0.5 * inter * (1.0 / na + 1.0 / nb))
 
 
 def tversky(a, b, *, alpha: float = 0.5, beta: float = 0.5) -> float:
